+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 Version next
 ============
-Released: 2020-01-13
+Released: 2020-02-16
 
  * Pascal+ workaround for rootless Xorg distributions (e.g. Fedora). More info
    on !45.
+ * Fix #101: Add notification dialog if proprietary NVIDIA driver not installed
 
 Version 0.13.3
 ==============